product is deleting and page is refreshing

+++ client/delete.js
@@ -1,18 +1,20 @@
 window.addEventListener("click", function(event) {
-    let productId = event.target.id;
-    console.log("product id", productId);
-    fetch(`/myproducts/${productId}`, {
-        method: 'delete',
-        credentials: "include"
-    })
-    .then( function(response) {
-        window.location.reload(true);
-        // console.log("response", response);
-        return response;
-    })
-    .catch( error => {
-        console.log("are you the JSON issue", error);
-    })
+    if (event.target.id > 0) {
+        let productId = event.target.id;
+        console.log("product id", productId);
+        fetch(`/myproducts/${productId}`, {
+            method: 'delete',
+            credentials: "include"
+        })
+        .then( function(response) {
+            document.location.reload(true);
+            console.log("response", response);
+            return response;
+        })
+        .catch( error => {
+            console.log("are you the JSON issue", error);
+        }) 
+    }
 });
